requirements.txt
pypdf
pandas
openpyxl

Office365-REST-Python-Client